[BE fix] DM 이전 채팅 데이터 페이지네이션

+++ server/src/main/java/com/ssafy/server/ScheduledTask.java
@@ -24,7 +24,7 @@ public class ScheduledTask {
                 String keyNumStr = keyName.replaceAll("[^0-9]", "");
                 chatService.saveToJPA(chatService.loadFromRedis(keyNumStr, false, true));
             }
-            else{
+            else if(keyName.startsWith("oldChat")){
                 chatService.deleteKeyInRedis(keyName);
             }
         }
